🎨 Add validate module
