+++ html/signup.html
@@ -32,13 +32,13 @@
           <div class="input-div">
             <label for="password">비밀번호</label>
             <input type="password" id="password" placeholder="비밀번호를 입력해주세요" />
-            <img src="/images/eye-off.svg" class="btn-visibility btn-visibility-pwd" />
+            <img src="/images/eye-off.svg" class="btn-visibility" />
             <div class="error-msg-pwd error-msg"></div>
           </div>
           <div class="input-div">
             <label for="password-repeat">비밀번호 확인</label>
             <input type="password" id="password-repeat" placeholder="비밀번호를 다시 한 번 입력해주세요" />
-            <img src="/images/eye-off.svg" class="btn-visibility btn-visibility-repeat" />
+            <img src="/images/eye-off.svg" class="btn-visibility" />
             <div class="error-msg-repeat error-msg"></div>
           </div>
 
@@ -66,6 +66,6 @@
         </div>
       </div>
     </main>
-    <script src="/js/signup.js"></script>
+    <script type="module" src="/js/signup.js"></script>
   </body>
 </html>
